Move ProductHeaderV2 to a folder
